# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My POS App API",
        "description": "API documentation for My POS App backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/purchase-invoice-payment-links": {
            "get": {
                "tags": [
                    "PurchaseInvoicePaymentLink"
                ],
                "summary": "List all purchase invoice payment links",
                "operationId": "7e6dfc0d1ba059150fe8bb42acc5ce54",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "PurchaseInvoicePaymentLink"
                ],
                "summary": "Create a new purchase invoice payment link",
                "operationId": "b0aeb39f11e5f4a630563f1d384c30d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase invoice payment link created"
                    }
                }
            }
        },
        "/purchase-invoice-payment-links/{id}": {
            "get": {
                "tags": [
                    "PurchaseInvoicePaymentLink"
                ],
                "summary": "Get a single purchase invoice payment link",
                "operationId": "2770f522bf4ebe10c4dd4ab7d982c2c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Purchase invoice payment link not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "PurchaseInvoicePaymentLink"
                ],
                "summary": "Update a purchase invoice payment link",
                "operationId": "66aa6cfc702a4b4f11eaae10b50fb9cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase invoice payment link updated"
                    },
                    "404": {
                        "description": "Purchase invoice payment link not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PurchaseInvoicePaymentLink"
                ],
                "summary": "Delete a purchase invoice payment link",
                "operationId": "f0043b80296b93475a2da538715485cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Purchase invoice payment link deleted"
                    },
                    "404": {
                        "description": "Purchase invoice payment link not found"
                    }
                }
            }
        },
        "/purchase-payments": {
            "get": {
                "tags": [
                    "PurchasePayment"
                ],
                "summary": "List all purchase payments",
                "operationId": "bef61895fcdaa48f58aad6564916c7dd",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "PurchasePayment"
                ],
                "summary": "Create a new purchase payment",
                "operationId": "a0fbe25411273064f22c33cce13420ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase payment created"
                    }
                }
            }
        },
        "/purchase-payments/{id}": {
            "get": {
                "tags": [
                    "PurchasePayment"
                ],
                "summary": "Get a single purchase payment",
                "operationId": "94cd43deb098701e408c8a6627ac87c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Purchase payment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "PurchasePayment"
                ],
                "summary": "Update a purchase payment",
                "operationId": "f6999d8cf5ebb9e96a1c04cf0cec9659",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase payment updated"
                    },
                    "404": {
                        "description": "Purchase payment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PurchasePayment"
                ],
                "summary": "Delete a purchase payment",
                "operationId": "089d6fa59a8faec367df20c5c6f84bf6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Purchase payment deleted"
                    },
                    "404": {
                        "description": "Purchase payment not found"
                    }
                }
            }
        },
        "/sales": {
            "get": {
                "tags": [
                    "Sale"
                ],
                "summary": "List all sales",
                "operationId": "757015b68ec0d83a8c347a33d3e0f0fe",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sale"
                ],
                "summary": "Create a new sale",
                "operationId": "ef46f4e9b7cc0d0b49e61654314dd5e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created"
                    }
                }
            }
        },
        "/sales/{id}": {
            "get": {
                "tags": [
                    "Sale"
                ],
                "summary": "Get a single sale",
                "operationId": "9fd425e864ae96a5ea2930d595205cdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sale"
                ],
                "summary": "Update a sale",
                "operationId": "d01798024447d3cab8cdd85801b9c5fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sale updated"
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sale"
                ],
                "summary": "Delete a sale",
                "operationId": "3fc376ab0dc0bc77c9b966838f9e8709",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sale deleted"
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                }
            }
        },
...
    },
...
}